Moved AIFProgressMonitor to its own Java file.  A little bit of cleanup.
